add functional of buying and selling the BTC & USDT (NOT COMPLETED

+++ Web-Api.online/Models/StoredProcedures/spGet_BTC_USDT_OpenOrdersResult.cs
@@ -7,9 +7,9 @@ namespace Web_Api.online.Models.StoredProcedures
         public long OpenOrderId { get; set; }
         public DateTime CreateDate { get; set; }
         public bool? IsBuy { get; set; }
-        public double Price { get; set; }
-        public double Amount { get; set; }
-        public double Total { get; set; }
+        public decimal Price { get; set; }
+        public decimal Amount { get; set; }
+        public decimal Total { get; set; }
         public string CreateUserId { get; set; }
     }
 }
